-moved Slider styles + variables to teams theme
