+++ packages/react/src/themes/base/componentVariables.ts
@@ -6,5 +6,3 @@ export { default as Icon } from './components/Icon/iconVariables'
 export { default as Loader } from './components/Loader/loaderVariables'
 
 export { default as Grid } from './components/Grid/gridVariables'
-
-export { default as Slider } from './components/Slider/sliderVariables'
